Extension/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
bottle==0.12.25
bottle-websocket==0.2.9
certifi==2023.5.7
charset-normalizer==3.1.0
dnspython==2.3.0
frozenlist==1.3.3
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.2
idna==3.4
multidict==6.0.4
openai==0.27.6
Paste==3.5.3
pymongo==4.3.3
python-dotenv==1.0.0
regex==2023.5.5
requests==2.30.0
six==1.16.0
tiktoken==0.4.0
tqdm==4.65.0
urllib3==2.0.2
yarl==1.9.2
youtube-transcript-api==0.6.0
zope.event==4.6
zope.interface==6.0
